--- a/addons/hr/models/res_users.py
+++ b/addons/hr/models/res_users.py
@@ -179,18 +179,9 @@ class User(models.Model):
         # Note: limit the `sudo` to the only action of "editing own profile" action in order to
         # avoid breaking `groups` mecanism on res.users form view.
         profile_view = self.env.ref("hr.res_users_view_form_profile")
-        original_user = self.env.user
         if profile_view and view_id == profile_view.id:
             self = self.with_user(SUPERUSER_ID)
         result = super(User, self).get_view(view_id, view_type, **options)
-        # Due to using the SUPERUSER the result will contain action that the user may not have access too
-        # here we filter out actions that requires special implicit rights to avoid having unusable actions
-        # in the dropdown menu.
-        if options.get('toolbar') and self.env.user != original_user:
-            self = self.with_user(original_user.id)
-            if not self.user_has_groups("base.group_erp_manager"):
-                change_password_action = self.env.ref("base.change_password_wizard_action")
-                result['toolbar']['action'] = [act for act in result['toolbar']['action'] if act['id'] != change_password_action.id]
         return result
 
     @api.model_create_multi