The master plan finally comes to an end.

Thanks to:
- odoo/odoo#87522 refactoring `load_views`,
- odoo/odoo#94337 refactoring `common.Form` to prevent changing
    invisible fields in unit tests using `Form` instances,
- odoo/odoo#95729 refactoring the behavior of `groups=` in views,
- odoo/odoo#98551 removing the need of the `groups_id` many2many field
    on back-end views.

The result returned by `get_view`/`get_views` can now finally be easily
and efficiently cached, in order to cache back-end views.

The goal of this revision is to cache the model views and fields
already post-processed for the web client
(with the modifiers, etc., already computed)
without group restriction.
Then, from this cached version, post-process group related features,
such as removing nodes restricted with a `groups=` attribute,
set the create/write button according to the user access rights to models, ...

Not including the groups in the cache key allows:
- to have less cached versions,
  (otherwise it would be one cached version per different group combination)
- to not have to fetch the user groups to compute the key
  (with the current cache key,
  there is nothing to fetch from the database to compute the key)

Besides, post-processing the groups features
after taking the view from the cache of the view doesn't take a tremendous time:
 - parsing arch from/to string with etree is fast,
 - removing the `groups=` nodes using etree is fast,
 - adding the `create="False"`, `write="False"`, `delete="False"` on the view
   root node according to the access right of the user on the model is fast.

This allows way faster calls to `get_views` by the web client,
as the server no longer need, for each call, to fetch the views in database,
combine the inherited views, post-process the modifiers attributes, etc.

Enterprise PR: odoo/enterprise#30974

From easiest case to worst case:

### res.company (small view)

#### only base installed, with list and form, no toolbar, no filter

Before
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form')]); self.env.invalidate_all();
CPU times: user 8.96 s, sys: 415 ms, total: 9.37 s
Wall time: 12.8 s
```
After
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form')]); self.env.invalidate_all();
CPU times: user 443 ms, sys: 0 ns, total: 443 ms
Wall time: 474 ms
```

#### only base installed, with list and form, with toolbar, no filter

Before
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form')], options={'toolbar': True}); self.env.invalidate_all();
CPU times: user 9.1 s, sys: 462 ms, total: 9.56 s
Wall time: 13.5 s
```
After
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form')], options={'toolbar': True}); self.env.invalidate_all();
CPU times: user 480 ms, sys: 117 µs, total: 480 ms
Wall time: 480 ms
```

#### only base installed, with list, form and search, with toolbar, with filters

Before
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form'), (False, 'search')], options={'toolbar': True, 'load_filters': True}); self.env.invalidate_all();
CPU times: user 10.8 s, sys: 686 ms, total: 11.5 s
Wall time: 16.3 s
```
After
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form'), (False, 'search')], options={'toolbar': True, 'load_filters': True}); self.env.invalidate_all();
CPU times: user 1.4 s, sys: 83.9 ms, total: 1.49 s
Wall time: 2.02 s
```

#### all modules installed, with list and form, no toolbar, no filter

Before
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form')]); self.env.invalidate_all();
CPU times: user 26.2 s, sys: 930 ms, total: 27.2 s
Wall time: 33.8 s
```
After
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form')]); self.env.invalidate_all();
CPU times: user 1.49 s, sys: 38.1 ms, total: 1.53 s
Wall time: 1.58 s
```

#### all modules installed, with list and form, with toolbar, no filter

Before
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form')], options={'toolbar': True}); self.env.invalidate_all();
CPU times: user 25.5 s, sys: 1.13 s, total: 26.7 s
Wall time: 34.4 s
```
After
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form')], options={'toolbar': True}); self.env.invalidate_all();
CPU times: user 1.55 s, sys: 30.9 ms, total: 1.58 s
Wall time: 1.64 s
```

#### all modules installed, with list, form and search, with toolbar with filters

Before
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form'), (False, 'search')], options={'toolbar': True, 'load_filters': True}); self.env.invalidate_all();
CPU times: user 29.6 s, sys: 1.29 s, total: 30.9 s
Wall time: 39.5 s
```
After
```py
%time for i in range(1000): self.env['res.company'].get_views([(False, 'list'), (False, 'form'), (False, 'search')], options={'toolbar': True, 'load_filters': True}); self.env.invalidate_all();
CPU times: user 2.65 s, sys: 48.8 ms, total: 2.69 s
Wall time: 3.22 s
```

### res.partner (big view, with get_view override)

#### only base installed, with list and form, no toolbar, no filter

Before
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form')]); self.env.invalidate_all();
CPU times: user 24.9 s, sys: 787 ms, total: 25.7 s
Wall time: 31.3 s
```
After
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form')]); self.env.invalidate_all();
CPU times: user 1.48 s, sys: 973 µs, total: 1.48 s
Wall time: 1.5 s
```

#### only base installed, with list and form, with toolbar, no filter

Before
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form')], options={'toolbar': True}); self.env.invalidate_all();
CPU times: user 25.8 s, sys: 729 ms, total: 26.5 s
Wall time: 32.2 s
```
After
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form')], options={'toolbar': True}); self.env.invalidate_all();
CPU times: user 1.5 s, sys: 0 ns, total: 1.5 s
Wall time: 1.5 s
```

#### only base installed, with list, form and search, with toolbar, with filters

Before
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form'), (False, 'search')], options={'toolbar': True, 'load_filters': True}); self.env.invalidate_all();
CPU times: user 30.2 s, sys: 1.13 s, total: 31.4 s
Wall time: 40.1 s
```
After
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form'), (False, 'search')], options={'toolbar': True, 'load_filters': True}); self.env.invalidate_all();
CPU times: user 2.8 s, sys: 61.2 ms, total: 2.86 s
Wall time: 3.53 s
```

#### all modules installed, with list and form, no toolbar, no filter

Before
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form')]); self.env.invalidate_all();
CPU times: user 1min 10s, sys: 1.15 s, total: 1min 11s
Wall time: 1min 21s
```
After
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form')]); self.env.invalidate_all();
CPU times: user 5.51 s, sys: 953 µs, total: 5.51 s
Wall time: 5.57 s
```

#### all modules installed, with list and form, with toolbar, no filter

Before
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form')], options={'toolbar': True}); self.env.invalidate_all();
CPU times: user 1min 12s, sys: 995 ms, total: 1min 13s
Wall time: 1min 24s
```
After
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form')], options={'toolbar': True}); self.env.invalidate_all();
CPU times: user 5.1 s, sys: 540 µs, total: 5.1 s
Wall time: 5.17 s
```

### all modules installed, with list, form and search, with toolbar, with filters

Before
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form'), (False, 'search')], options={'toolbar': True, 'load_filters': True}); self.env.invalidate_all();
CPU times: user 1min 22s, sys: 1.8 s, total: 1min 23s
Wall time: 1min 38s
```
After
```py
%time for i in range(1000): self.env['res.partner'].get_views([(False, 'list'), (False, 'form'), (False, 'search')], options={'toolbar': True, 'load_filters': True}); self.env.invalidate_all();
CPU times: user 5.6 s, sys: 99 ms, total: 5.7 s
Wall time: 6.36 s
```