@beledouxdenis 2022-09-09
> it seems this causes issues with `--dev xml`

This is most-likely true. I will have a look right now.