@mshahbazi 2022-09-09
a `conditional` decorator might address the issue. but I failed to make it work, since `tools.config` is empty at the stage of loading `models.py`