@beledouxdenis 2022-09-09
> a `conditional` decorator might address the issue. but I failed to make it work, since `tools.config` is empty at the stage of loading `models.py`

Yup, that is what I also tried. But for the reason you mention, it doesn't work.

I checked if it was doable to change the code so the loading of the config is done prior to the loading of `models.py`,
but it's really not that easy.

I will try to move all these `get_view` method to the module `web` instead of `base`: It will overcome the issue as in the same time make things cleaner, because these `get_view` methods do not make sense without the web client, which is in the module `web`, not `base`. I already planned to do that at some point, but it looks like I will be forced earlier than expected, for this.