@beledouxdenis 2022-10-05
> you are right. moving the code to the `web` makes more sense. I would be happy to be of any help. however, for no reason, every commit I make, gets red! I don't know why :|
> 
> btw, thanks you so much for the great work.

The issue is now solved thanks to odoo/odoo#101200 :).

Thank you for the kind words !